+++ CHANGELOG.md
@@ -9,6 +9,16 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+## [7.6.4] - 2026-04-28
+
+### Added
+
+- special characters popup used in console
+
+### Fixed
+
+- **autosave**: handle nils in `autosave_info_component`
+
 ## [7.6.3] - 2026-04-22
 
 ### Fixed